backend/requirement.txt
asgiref==3.10.0
contourpy==1.3.3
coverage==7.12.0
cycler==0.12.1
Django==5.2.7
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.61.0
iniconfig==2.3.0
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.5
packaging==25.0
Pillow==10.1.0
pluggy==1.6.0
psycopg2-binary==2.9.9
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pytest==9.0.1
pytest-cov==7.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
wheel==0.45.1
